feat: popup for http is not allowed
