+++ vite-pwa.config.ts
@@ -1,5 +1,6 @@
 import { mergeConfig, defineConfig } from 'vite'
 import { VitePWA } from 'vite-plugin-pwa'
+import basicSsl from '@vitejs/plugin-basic-ssl'
 
 import viteBaseConfig from './vite-base.config'
 import { manifest } from './vite-config/manifest'
@@ -21,7 +22,8 @@ export default mergeConfig(
           globPatterns: ['**/*.{js,css,html,ico,png,svg,wasm}'],
           maximumFileSizeToCacheInBytes: 5000000 // 5 MiB
         }
-      })
+      }),
+      basicSsl()
     ]
   })
 )
